feat: offer deployment manifests
Adds a basic Dockerfile for building the image containing this
application.
Additionally adds k8s manifest files that allow for deploying
the application to a Kubernetes cluster. Alongside with that
a deployment script is introduced that can be used to deploy
the application to an existing minikube instance for development
purposes.