feat: add message subscription endpoint
Adds an HTTP endpoint for registering a message subscription
for a specific analysis using a webhook URL. This allows for
clients being informed about new messages without the need to
periodicaly poll for them (increasing load on the server).

Also offers an endpoint for getting back information regarding
the message subscription.

Subscriptions are saved using an underlying MongoDB instance.