feat: forward received messages to webhook
Forwards any received message of an anlysis to a client
that has subscribed for it. If there is no subscription
then messages will be silently dropped.
Allows for message forwarding using HTTP and HTTPS.
However, in case of HTTPS there is currently no way to
configure TLS using self-signed certificates. This can
be implemented later on.