feat: offer discovery endpoint
Adds an endpoint that can be used to discover participating
nodes within an analysis.
This functionality is backed by the central side (hub) from
which this information is gathered. However, availability of
the central side is not integrated into the health checks
since requests to downstream services can simply lead to
appropriate HTTP status codes (502).

For future development one could still try to integrate some
kind of circuit breaker though.