feat: revise message subscription endpoint
Adjusts the endpoint used for message subscriptions. Bases it off of
an analysis which moves the analysis identifier from the request body
to the actual URL.
This has been done to put an emphasis on the fact that subscriptions
for messages are grouped by an analysis identifier.

Also adds existence checks before adding a message subscription. When
an analysis does not exist on the central side (hub) then the client
must not be able to add a message subscription.
This is done because there won't be any incoming messages if for an
analysis if it does not exist in the first place.