feat: revise message sending functionality
Adds functionality to send messages to specific recipients
instead of using a broadcast. This is expected to be used
in conjunction with the discovery endpoints.
Also guards the recipients in a way such that only messages
to participating nodes within an analysis context are valid.
When trying to send messages to other participants then this
will lead to an error.

Furthermore adjusts the broadcast endpoint. This change is
made since everything is dependant upon an analysis in the
first place and this change makes that clear.