fix: reference valid artifact in dockerfile
Adjusts the artifact that is being referenced by the Dockerfile.
Uses an .cjs file since the message broker is using nestjs which
itself makes use of express. The latter being a CommonJS module.
To avoid any JS shenanigans when it comes to correctly importing
dependencies it was decided to go with the .cjs file.