feat: use effect with express integration
This initiates the start for transitioning from nestjs to
effect with a pure express integration. This is mainly done
to leverage features provided by effect for better error
management.

For this a simple express server is started bootstrapped
with a basic router that does not yet pick up all the routes
that are already there (from the previous nestjs implementation).

Signed-off-by: Alexander Twrdik <6052859+DiCanio@users.noreply.github.com>