Require Authentication for all Requests
All requests are required to be authenticated. This also includes
health endpoints (will be changed later on).
Thus, a client needs to provide a valid JWT in order for the
request to get through. Validation of this token is accomplished
by leveraging an OpenID provider using its JWKS URL. Via the same
URL the application automatically runs discovery processes regarding
the provider.

Signed-off-by: Alexander Twrdik <6052859+DiCanio@users.noreply.github.com>