Implement Sending Unencrypted Messages
Adds endpoints for sending messages to specific participants
of an analysis or for broadcasting a message to all of them.
Messages are sent using socket.io. Security checks are not
in place yet and will be added with a later commit. Also,
messages are sent unencrypted. Encryption will also be added
in a later commit.

Signed-off-by: Alexander Twrdik <6052859+DiCanio@users.noreply.github.com>