Adjust Message Handling with Socket.IO
This takes into account that the java client for Socket.IO
mandates which serialization library should be used. This
is however, only true when sending messages and not when
receiving them. In the latter case every arbitrary library
for deserializing a raw JSON byte stream can be used.
Thus, for sending adjustments have been made to make use
of 'org.json' as Socket.IO's choice of a serialization lib.
If there's an overlap / reuse with the part that deals with
receiving messages then also jackson is incorporated at the
same time. This decision has been made since jackson offers
more control over the deserialization process compared to
org.json.

At the same time some architectural changes are made which
should help with testing later on.

Signed-off-by: Alexander Twrdik <6052859+DiCanio@users.noreply.github.com>