Add E2E-Encryption for Messages
Adds end-to-end encryption for messages sent via
the broker.
Encryption/Decryption is done via a derived
symmetric key in combination with AES-256 bit
and GCM without any padding. In order for the
other party to decrypt messages, initialization
vector information of prepended to encrypted
messages and becomes part of the actual message.

Changes require an ECDH private key to be
provided during startup. Required public keys
of the receiving parties are dynamically
fetched from the Hub (although not yet cached).

Signed-off-by: Alexander Twrdik <6052859+DiCanio@users.noreply.github.com>