Authorize All Requests to Health Endpoint
Makes sure that all requests to the health endpoint
(/actuator/health) are permitted without the need
of having a valid JWT. This is especially necessary
for running the healthcheck that's set up within the
corresponding Dockerfile.

Keeps other actuator endpoints untouched so that for
these a valid JWT is still required. This is done
because some of them allow access to potentially
sensitive information (e.g. logfiles etc.).

Signed-off-by: Alexander Twrdik <6052859+DiCanio@users.noreply.github.com>