Read Private Key From File
Expects the node's private key to be provisioned as
a file instead as an environment variable. This is
done so that Docker secrets can be properly used
when setting up the application in a docker-compose
environment. Also, this way secrets are not
polluting environment variables for which there are
no security checks in place.

Signed-off-by: Alexander Twrdik <6052859+DiCanio@users.noreply.github.com>