Read Robot Secret From File
Expects the node's associated robot account secret
to be provisioned as a file instead of an environment
variable. This is done so that Docker secrets can be
properly used when setting up the application in a
docker-compose environment. Also, this way secrets
are not polluting environment variables for which
there are no security checks in place.

Signed-off-by: Alexander Twrdik <6052859+DiCanio@users.noreply.github.com>