Ensure Robot Secret Getting Base64 Decoded
This is a follow-up for #73eee13e that ensures the
robot secret gets base64 decoded after being read.
This needs to be done when reading files in a k8s
environment since secrets are base64 encoded by
default. Thus, the whole application expects the
robot secret to be given being base64 encoded.

Signed-off-by: Alexander Twrdik <6052859+DiCanio@users.noreply.github.com>