Add Proxy Configuration Options
Adds options for configuring forward proxy functionality
using HTTP (SOCKS is not supported by these changes).
Allows for setting the following options:
- hostname
- port
- username
- password (via file reference)
- whitelist

It comes with support for hostname whitelisting, effectively
bypassing the proxy if the requested host matches a pattern
(given as a Java RegEx).

These changes contain 2 different proxy configuration paths
which are necessary since the HTTP client used throughout the
application is different from the one used exclusively for
socket.io connections. The configuration paths provide the
same functionality but have to deal with different types of
clients that are configured in different ways.