Map NodeId to RobotId When Emitting Messages
Ensures that node IDs are used for specifying message
receivers and maps them internally to a robot ID.
Mapping this information is crucial since the robot
ID is used for the underlying communication via a
socket and is not known to the client. The node ID
however, is well known to the client as it can get
fetched using discovery endpoints.